+++ sdk/src/main/java/jp/co/freee/accounting/models/InvoiceCreateParamsInvoiceContents.java
@@ -58,15 +58,15 @@ public class InvoiceCreateParamsInvoiceContents {
 
   public static final String SERIALIZED_NAME_SEGMENT1_TAG_ID = "segment_1_tag_id";
   @SerializedName(SERIALIZED_NAME_SEGMENT1_TAG_ID)
-  private Integer segment1TagId;
+  private Long segment1TagId;
 
   public static final String SERIALIZED_NAME_SEGMENT2_TAG_ID = "segment_2_tag_id";
   @SerializedName(SERIALIZED_NAME_SEGMENT2_TAG_ID)
-  private Integer segment2TagId;
+  private Long segment2TagId;
 
   public static final String SERIALIZED_NAME_SEGMENT3_TAG_ID = "segment_3_tag_id";
   @SerializedName(SERIALIZED_NAME_SEGMENT3_TAG_ID)
-  private Integer segment3TagId;
+  private Long segment3TagId;
 
   public static final String SERIALIZED_NAME_TAG_IDS = "tag_ids";
   @SerializedName(SERIALIZED_NAME_TAG_IDS)
@@ -289,7 +289,7 @@ public class InvoiceCreateParamsInvoiceContents {
   }
 
 
-  public InvoiceCreateParamsInvoiceContents segment1TagId(Integer segment1TagId) {
+  public InvoiceCreateParamsInvoiceContents segment1TagId(Long segment1TagId) {
     
     this.segment1TagId = segment1TagId;
     return this;
@@ -304,17 +304,17 @@ public class InvoiceCreateParamsInvoiceContents {
   @javax.annotation.Nullable
   @ApiModelProperty(example = "1", value = "セグメント１ID")
 
-  public Integer getSegment1TagId() {
+  public Long getSegment1TagId() {
     return segment1TagId;
   }
 
 
-  public void setSegment1TagId(Integer segment1TagId) {
+  public void setSegment1TagId(Long segment1TagId) {
     this.segment1TagId = segment1TagId;
   }
 
 
-  public InvoiceCreateParamsInvoiceContents segment2TagId(Integer segment2TagId) {
+  public InvoiceCreateParamsInvoiceContents segment2TagId(Long segment2TagId) {
     
     this.segment2TagId = segment2TagId;
     return this;
@@ -329,17 +329,17 @@ public class InvoiceCreateParamsInvoiceContents {
   @javax.annotation.Nullable
   @ApiModelProperty(example = "1", value = "セグメント２ID")
 
-  public Integer getSegment2TagId() {
+  public Long getSegment2TagId() {
     return segment2TagId;
   }
 
 
-  public void setSegment2TagId(Integer segment2TagId) {
+  public void setSegment2TagId(Long segment2TagId) {
     this.segment2TagId = segment2TagId;
   }
 
 
-  public InvoiceCreateParamsInvoiceContents segment3TagId(Integer segment3TagId) {
+  public InvoiceCreateParamsInvoiceContents segment3TagId(Long segment3TagId) {
     
     this.segment3TagId = segment3TagId;
     return this;
@@ -354,12 +354,12 @@ public class InvoiceCreateParamsInvoiceContents {
   @javax.annotation.Nullable
   @ApiModelProperty(example = "1", value = "セグメント３ID")
 
-  public Integer getSegment3TagId() {
+  public Long getSegment3TagId() {
     return segment3TagId;
   }
 
 
-  public void setSegment3TagId(Integer segment3TagId) {
+  public void setSegment3TagId(Long segment3TagId) {
     this.segment3TagId = segment3TagId;
   }
