+++ sdk/src/main/java/jp/co/freee/accounting/models/InvoiceInvoiceContents.java
@@ -42,7 +42,7 @@ public class InvoiceInvoiceContents {
 
   public static final String SERIALIZED_NAME_AMOUNT = "amount";
   @SerializedName(SERIALIZED_NAME_AMOUNT)
-  private Integer amount;
+  private Long amount;
 
   public static final String SERIALIZED_NAME_DESCRIPTION = "description";
   @SerializedName(SERIALIZED_NAME_DESCRIPTION)
@@ -82,7 +82,7 @@ public class InvoiceInvoiceContents {
 
   public static final String SERIALIZED_NAME_SEGMENT1_TAG_ID = "segment_1_tag_id";
   @SerializedName(SERIALIZED_NAME_SEGMENT1_TAG_ID)
-  private Integer segment1TagId;
+  private Long segment1TagId;
 
   public static final String SERIALIZED_NAME_SEGMENT1_TAG_NAME = "segment_1_tag_name";
   @SerializedName(SERIALIZED_NAME_SEGMENT1_TAG_NAME)
@@ -90,7 +90,7 @@ public class InvoiceInvoiceContents {
 
   public static final String SERIALIZED_NAME_SEGMENT2_TAG_ID = "segment_2_tag_id";
   @SerializedName(SERIALIZED_NAME_SEGMENT2_TAG_ID)
-  private Integer segment2TagId;
+  private Long segment2TagId;
 
   public static final String SERIALIZED_NAME_SEGMENT2_TAG_NAME = "segment_2_tag_name";
   @SerializedName(SERIALIZED_NAME_SEGMENT2_TAG_NAME)
@@ -98,7 +98,7 @@ public class InvoiceInvoiceContents {
 
   public static final String SERIALIZED_NAME_SEGMENT3_TAG_ID = "segment_3_tag_id";
   @SerializedName(SERIALIZED_NAME_SEGMENT3_TAG_ID)
-  private Integer segment3TagId;
+  private Long segment3TagId;
 
   public static final String SERIALIZED_NAME_SEGMENT3_TAG_NAME = "segment_3_tag_name";
   @SerializedName(SERIALIZED_NAME_SEGMENT3_TAG_NAME)
@@ -230,7 +230,7 @@ public class InvoiceInvoiceContents {
   }
 
 
-  public InvoiceInvoiceContents amount(Integer amount) {
+  public InvoiceInvoiceContents amount(Long amount) {
     
     this.amount = amount;
     return this;
@@ -244,12 +244,12 @@ public class InvoiceInvoiceContents {
   **/
   @ApiModelProperty(example = "108000", required = true, value = "金額")
 
-  public Integer getAmount() {
+  public Long getAmount() {
     return amount;
   }
 
 
-  public void setAmount(Integer amount) {
+  public void setAmount(Long amount) {
     this.amount = amount;
   }
 
@@ -466,7 +466,7 @@ public class InvoiceInvoiceContents {
   }
 
 
-  public InvoiceInvoiceContents segment1TagId(Integer segment1TagId) {
+  public InvoiceInvoiceContents segment1TagId(Long segment1TagId) {
     
     this.segment1TagId = segment1TagId;
     return this;
@@ -481,12 +481,12 @@ public class InvoiceInvoiceContents {
   @javax.annotation.Nullable
   @ApiModelProperty(example = "1", value = "セグメント１ID")
 
-  public Integer getSegment1TagId() {
+  public Long getSegment1TagId() {
     return segment1TagId;
   }
 
 
-  public void setSegment1TagId(Integer segment1TagId) {
+  public void setSegment1TagId(Long segment1TagId) {
     this.segment1TagId = segment1TagId;
   }
 
@@ -514,7 +514,7 @@ public class InvoiceInvoiceContents {
   }
 
 
-  public InvoiceInvoiceContents segment2TagId(Integer segment2TagId) {
+  public InvoiceInvoiceContents segment2TagId(Long segment2TagId) {
     
     this.segment2TagId = segment2TagId;
     return this;
@@ -529,12 +529,12 @@ public class InvoiceInvoiceContents {
   @javax.annotation.Nullable
   @ApiModelProperty(example = "1", value = "セグメント２ID")
 
-  public Integer getSegment2TagId() {
+  public Long getSegment2TagId() {
     return segment2TagId;
   }
 
 
-  public void setSegment2TagId(Integer segment2TagId) {
+  public void setSegment2TagId(Long segment2TagId) {
     this.segment2TagId = segment2TagId;
   }
 
@@ -562,7 +562,7 @@ public class InvoiceInvoiceContents {
   }
 
 
-  public InvoiceInvoiceContents segment3TagId(Integer segment3TagId) {
+  public InvoiceInvoiceContents segment3TagId(Long segment3TagId) {
     
     this.segment3TagId = segment3TagId;
     return this;
@@ -577,12 +577,12 @@ public class InvoiceInvoiceContents {
   @javax.annotation.Nullable
   @ApiModelProperty(example = "1", value = "セグメント３ID")
 
-  public Integer getSegment3TagId() {
+  public Long getSegment3TagId() {
     return segment3TagId;
   }
 
 
-  public void setSegment3TagId(Integer segment3TagId) {
+  public void setSegment3TagId(Long segment3TagId) {
     this.segment3TagId = segment3TagId;
   }
