+++ sdk/src/main/java/jp/co/freee/accounting/models/ManualJournalDetails.java
@@ -96,7 +96,7 @@ public class ManualJournalDetails {
 
   public static final String SERIALIZED_NAME_ID = "id";
   @SerializedName(SERIALIZED_NAME_ID)
-  private Integer id;
+  private Long id;
 
   public static final String SERIALIZED_NAME_ITEM_ID = "item_id";
   @SerializedName(SERIALIZED_NAME_ITEM_ID)
@@ -132,7 +132,7 @@ public class ManualJournalDetails {
 
   public static final String SERIALIZED_NAME_SEGMENT1_TAG_ID = "segment_1_tag_id";
   @SerializedName(SERIALIZED_NAME_SEGMENT1_TAG_ID)
-  private Integer segment1TagId;
+  private Long segment1TagId;
 
   public static final String SERIALIZED_NAME_SEGMENT1_TAG_NAME = "segment_1_tag_name";
   @SerializedName(SERIALIZED_NAME_SEGMENT1_TAG_NAME)
@@ -140,7 +140,7 @@ public class ManualJournalDetails {
 
   public static final String SERIALIZED_NAME_SEGMENT2_TAG_ID = "segment_2_tag_id";
   @SerializedName(SERIALIZED_NAME_SEGMENT2_TAG_ID)
-  private Integer segment2TagId;
+  private Long segment2TagId;
 
   public static final String SERIALIZED_NAME_SEGMENT2_TAG_NAME = "segment_2_tag_name";
   @SerializedName(SERIALIZED_NAME_SEGMENT2_TAG_NAME)
@@ -148,7 +148,7 @@ public class ManualJournalDetails {
 
   public static final String SERIALIZED_NAME_SEGMENT3_TAG_ID = "segment_3_tag_id";
   @SerializedName(SERIALIZED_NAME_SEGMENT3_TAG_ID)
-  private Integer segment3TagId;
+  private Long segment3TagId;
 
   public static final String SERIALIZED_NAME_SEGMENT3_TAG_NAME = "segment_3_tag_name";
   @SerializedName(SERIALIZED_NAME_SEGMENT3_TAG_NAME)
@@ -263,7 +263,7 @@ public class ManualJournalDetails {
   }
 
 
-  public ManualJournalDetails id(Integer id) {
+  public ManualJournalDetails id(Long id) {
     
     this.id = id;
     return this;
@@ -277,12 +277,12 @@ public class ManualJournalDetails {
   **/
   @ApiModelProperty(example = "1", required = true, value = "貸借行ID")
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
 
@@ -477,7 +477,7 @@ public class ManualJournalDetails {
   }
 
 
-  public ManualJournalDetails segment1TagId(Integer segment1TagId) {
+  public ManualJournalDetails segment1TagId(Long segment1TagId) {
     
     this.segment1TagId = segment1TagId;
     return this;
@@ -492,12 +492,12 @@ public class ManualJournalDetails {
   @javax.annotation.Nullable
   @ApiModelProperty(example = "1", value = "セグメント１ID")
 
-  public Integer getSegment1TagId() {
+  public Long getSegment1TagId() {
     return segment1TagId;
   }
 
 
-  public void setSegment1TagId(Integer segment1TagId) {
+  public void setSegment1TagId(Long segment1TagId) {
     this.segment1TagId = segment1TagId;
   }
 
@@ -525,7 +525,7 @@ public class ManualJournalDetails {
   }
 
 
-  public ManualJournalDetails segment2TagId(Integer segment2TagId) {
+  public ManualJournalDetails segment2TagId(Long segment2TagId) {
     
     this.segment2TagId = segment2TagId;
     return this;
@@ -540,12 +540,12 @@ public class ManualJournalDetails {
   @javax.annotation.Nullable
   @ApiModelProperty(example = "1", value = "セグメント２ID")
 
-  public Integer getSegment2TagId() {
+  public Long getSegment2TagId() {
     return segment2TagId;
   }
 
 
-  public void setSegment2TagId(Integer segment2TagId) {
+  public void setSegment2TagId(Long segment2TagId) {
     this.segment2TagId = segment2TagId;
   }
 
@@ -573,7 +573,7 @@ public class ManualJournalDetails {
   }
 
 
-  public ManualJournalDetails segment3TagId(Integer segment3TagId) {
+  public ManualJournalDetails segment3TagId(Long segment3TagId) {
     
     this.segment3TagId = segment3TagId;
     return this;
@@ -588,12 +588,12 @@ public class ManualJournalDetails {
   @javax.annotation.Nullable
   @ApiModelProperty(example = "1", value = "セグメント３ID")
 
-  public Integer getSegment3TagId() {
+  public Long getSegment3TagId() {
     return segment3TagId;
   }
 
 
-  public void setSegment3TagId(Integer segment3TagId) {
+  public void setSegment3TagId(Long segment3TagId) {
     this.segment3TagId = segment3TagId;
   }
